split hybrid_minmax.cpp
